Solving an NPV problem using Python code
Your firm is considering the launch of a new product, the XJ5. The upfront development cost
is $10 million, and you expect to earn a cash flow of $3 million per year for the next five years.
Plot the NPV profile for this project for discount rates ranging from 0% to 30%. For what
range of discount rates is the project attractive?